+++ smart-contracts/contracts/cl-vault/src/test_tube/proptest.rs
@@ -316,6 +316,7 @@ mod tests {
                         max_slippage: Decimal::bps(5), // optimize and check how this fits in the strategy as it could trigger organic errors we dont want to test
                         ratio_of_swappable_funds_to_use: Decimal::one(),
                         twap_window_seconds: 45,
+                        claim_after: None,
                     },
                 )),
                 &[],
