+++ smart-contracts/contracts/cl-vault/src/test_tube/range.rs
@@ -135,6 +135,7 @@ mod test {
                         max_slippage: Decimal::bps(9500),
                         ratio_of_swappable_funds_to_use: Decimal::one(),
                         twap_window_seconds: 45,
+                        claim_after: None,
                     },
                 )),
                 &[],
@@ -250,6 +251,7 @@ mod test {
                         max_slippage: Decimal::bps(9500),
                         ratio_of_swappable_funds_to_use: Decimal::one(),
                         twap_window_seconds: 45,
+                        claim_after: None,
                     },
                 )),
                 &[],
