+++ smart-contracts/contracts/cl-vault/src/vault/concentrated_liquidity.rs
@@ -201,7 +201,14 @@ mod tests {
 
         let position_id = 1;
         POSITION
-            .save(deps.as_mut().storage, &Position { position_id })
+            .save(
+                deps.as_mut().storage,
+                &Position {
+                    position_id,
+                    join_time: 0,
+                    claim_after: None,
+                },
+            )
             .unwrap();
 
         let result = withdraw_from_position(&mut deps.storage, &env, liquidity_amount).unwrap();
